Fixing the build.
CustomerMaintenanceController:  Added some error checking when user clicks on modify button.  Added new GetCustomerInfo method that will regrab the customer based on the id.

ModifyCustomerController:  Added some error checking to OnAcceptButtonClicked if the customer was not able to successfully be updated in the database.

CustomerMaintenanceForm:  Added functionality to clear the controls and customer, as well as the ability to grab the customer information from the database.

ModifyCustomerForm:  Now implements the correct interface.  Reworking to send old and new customers to controller to update the customer in the database.

ProgramFlowManager:  Now handling the case if there was an error while trying to modify the customer.